+++ lib/src/services/image_picker_service.dart
@@ -2,9 +2,9 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-import 'dart:typed_data';
-import 'package:flutter_image_picker/src/models/image_picker_config.dart';
-import 'package:image_picker/image_picker.dart';
+import "dart:typed_data";
+import "package:flutter_image_picker/src/models/image_picker_config.dart";
+import "package:image_picker/image_picker.dart";
 
 /// The Image Picker Service class is the functionality of the Image Picker
 /// package which uses the Image Picker package to choose an image.