+++ test/mocks/image_picker_service_mock.dart
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-import 'package:flutter_image_picker/src/services/image_picker_service.dart';
-import 'package:mocktail/mocktail.dart';
+import "package:flutter_image_picker/src/services/image_picker_service.dart";
+import "package:mocktail/mocktail.dart";
 
 class ImagePickerServiceMock extends Mock implements ImagePickerService {}